@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { FormBuilder, FormGroupDirective } from '@angular/forms';
+import { UntypedFormBuilder, FormGroupDirective } from '@angular/forms';
 import { ButtonComponent } from '@modules/buttons/components/button/button.component';
 
 import { ItemsListComponent } from './items-list.component';
@@ -11,7 +11,7 @@ describe('ItemsListComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [ItemsListComponent, ButtonComponent],
-      providers: [FormGroupDirective, FormBuilder],
+      providers: [FormGroupDirective, UntypedFormBuilder],
     }).compileComponents();
   });
 
